requirements.txt
certifi==2020.12.5
chardet==4.0.0
docutils==0.16
docxcompose==1.3.1
idna==2.10
Kivy==2.0.0
Kivy-Garden==0.1.4
lxml==4.6.2
Pillow==8.1.1
Pygments==2.8.0
python-docx==0.8.10
requests==2.25.1
six==1.15.0
urllib3==1.26.3
